+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## CHANGELOG
 
-1. hosted:video audio synced on peek, play, pause, stalled, waiting, canPlayThrough.
+1. rich:video support through react-player.